+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to Work Launcher will be documented in this file.
 
+## [1.0.20] - 2026-02-09
+
+### Added
+- Export Notes: Save all your notes to a JSON file for backup
+- Import Notes: Import notes from a previously exported JSON file
+- Download as PDF: Export the current note as a beautifully formatted PDF document
+
 ## [1.0.19] - 2026-02-08
 
 ### Changed